+++ buildSrc/src/main/java/net/neoforged/neodev/ProvideCompatibilityCheckJar.java
@@ -1,4 +1,4 @@
-package net.neoforged.neodev.jcc;
+package net.neoforged.neodev;
 
 import net.neoforged.neodev.e2e.InstallProductionClient;
 import org.gradle.api.GradleException;
